Fix models, add new endpoints

+++ src/routes/user.routes.js
@@ -6,25 +6,29 @@ const router = Router();
 
 // Consultar a todos los usuarios
 router.get('/users', async (req, res) => {
-    const users = await prisma.user.findMany();
-    res.json(users);
+    try {
+        const users = await prisma.user.findMany();
+        res.status(200).json(users);
+    } catch (error) {
+        console.error('Error al encontrar los usuarios:', error);
+        res.status(500).send('Error interno del servidor');
+    }
 });
 
 // Crear un nuevo usuario
 router.post('/user', async (req, res) => {
     try {
-        const { email, password } = req.body;
+        const { email, password, roleId } = req.body;
 
         // Verificar si el email ya está en uso
-        const existingUser = await prisma.user.findUnique({
+        const existingUserEmail = await prisma.user.findUnique({
             where: { email },
         });
 
-        if (existingUser) {
+        if (existingUserEmail) {
             return res.status(400).json({ message: 'El email ya está en uso' });
         }
 
-        // Hash de la contraseña
         const hashedPassword = await bcrypt.hash(password, 10);
 
         // Crear el nuevo usuario
@@ -32,12 +36,32 @@ router.post('/user', async (req, res) => {
             data: {
                 email,
                 password: hashedPassword,
+                roles: {
+                    create: {
+                        role: {
+                            connect: {
+                                id: roleId
+                            }
+                        }
+                    }
+                }
             },
+
+            // Muestra en el json informacion sobre la tabla pivote
+            include: {
+                roles: {
+                    include: {
+                        role: true
+                    }
+                }
+            }
         });
 
-        res.json(newUser);
+        res.status(201).json(newUser);
+
     } catch (error) {
-        res.status(500).json({ message: 'Error al crear el usuario', error });
+        console.error('Error al crear usuario:', error);
+        res.status(500).send('Error interno del servidor');
     }
 });
 
@@ -48,15 +72,25 @@ router.get('/user/:id', async (req, res) => {
 
         const user = await prisma.user.findUnique({
             where: { id: parseInt(id) },
+
+            // Muestra en el json informacion sobre la tabla pivote
+            include: {
+                roles: {
+                    include: {
+                        role: true
+                    }
+                }
+            }
         });
 
         if (!user) {
             return res.status(404).json({ message: 'Usuario no encontrado' });
         }
 
-        res.json(user);
+        res.status(200).json(user);
     } catch (error) {
-        res.status(500).json({ message: 'Error al obtener el usuario', error });
+        console.error('Error al obtener usuario:', error);
+        res.status(500).send('Error interno del servidor');
     }
 });
 
@@ -64,16 +98,75 @@ router.get('/user/:id', async (req, res) => {
 router.put('/user/:id', async (req, res) => {
     try {
         const { id } = req.params;
-        const { email, password } = req.body;
+        const { email, password, roleId } = req.body;
+
+        // Verificar si el email ya está en uso
+        if (email) {
+            
+            const existingUserEmail = await prisma.user.findUnique({
+                where: { email },
+            });
+
+            if (existingUserEmail) {
+                return res.status(400).json({ message: 'El email ya está en uso' });
+            }
+        }
+
+        // Re-encriptar la contraseña si se proporciona
+        const hashedPassword = password ? await bcrypt.hash(password, 10) : undefined;
+
+        // Actualizar el rol si se proporciona
+        if (roleId) {
+            const existingRole = await prisma.role.findUnique({
+                where: { id: roleId },
+            });
+
+            if (!existingRole) {
+                return res.status(404).json({ message: 'El rol proporcionado no existe.' });
+            }
+
+            // Eliminar todos los roles anteriores del usuario y asignar el nuevo rol
+            await prisma.usersOnRoles.deleteMany({
+                where: { userId: parseInt(id) }
+            });
+
+            await prisma.usersOnRoles.create({
+                data: {
+                    role: {
+                        connect: {
+                            id: roleId
+                        }
+                    },
+                    user: {
+                        connect: {
+                            id: parseInt(id)
+                        }
+                    }
+                }
+            });
+        }
 
         const updatedUser = await prisma.user.update({
-            where: { id: parseInt(id) },
-            data: { email, password },
+            where: {
+                id: parseInt(id)
+            },
+            data: {
+                email,
+                password: hashedPassword,
+            },
+            include: {
+                roles: {
+                    include: {
+                        role: true
+                    }
+                }
+            }
         });
 
-        res.json(updatedUser);
+        res.status(200).json(updatedUser);
     } catch (error) {
-        res.status(500).json({ message: 'Error al actualizar el usuario', error });
+        console.error('Error al actualizar usuario:', error);
+        res.status(500).send('Error interno del servidor');
     }
 });
 
@@ -82,6 +175,22 @@ router.delete('/user/:id', async (req, res) => {
     try {
         const { id } = req.params;
 
+        // Verificar si el usuario existe
+        const existingUser = await prisma.user.findUnique({
+            where: { id: parseInt(id) },
+        });
+
+        // Si el usuario no existe, retornar un mensaje de error
+        if (!existingUser) {
+            return res.status(404).json({ message: 'Usuario no encontrado' });
+        }
+
+        // Eliminar todas las relaciones del usuario en la tabla pivote
+        await prisma.usersOnRoles.deleteMany({
+            where: { userId: parseInt(id) },
+        });
+
+        // Eliminar el usuario
         await prisma.user.delete({
             where: { id: parseInt(id) },
         });
@@ -91,6 +200,5 @@ router.delete('/user/:id', async (req, res) => {
         res.status(500).json({ message: 'Error al eliminar el usuario', error });
     }
 });
-
 
 export default router;
